fix(Canvas) Clear error boundary after builds
