+++ editor/src/components/canvas/ui-jsx-canvas.tsx
@@ -1413,65 +1413,3 @@ const SpyWrapper: React.FunctionComponent<SpyWrapperProps> = (props) => {
     passThroughProps,
   )
 }
-
-interface CanvasErrorBoundaryProps extends CanvasReactReportErrorCallback {
-  filePath: string
-  fileCode: string
-}
-
-function isErrorObject(e: unknown): e is Error {
-  return typeof e === 'object' && e != null && 'name' in e && 'message' in e
-}
-
-function asErrorObject(e: unknown): Error {
-  // Required because JS supports throwing anything at all
-  if (isErrorObject(e)) {
-    return e
-  } else if (typeof e === 'string') {
-    return new Error(e)
-  } else {
-    return new Error(JSON.stringify(e))
-  }
-}
-
-interface CanvasErrorBoundaryState {
-  hasError: boolean
-}
-
-export class CanvasErrorBoundary extends React.Component<
-  CanvasErrorBoundaryProps,
-  CanvasErrorBoundaryState
-> {
-  constructor(props: CanvasErrorBoundaryProps) {
-    super(props)
-    this.state = { hasError: false }
-  }
-
-  static getDerivedStateFromError(_error: unknown): CanvasErrorBoundaryState {
-    return {
-      hasError: true,
-    }
-  }
-
-  componentDidUpdate(
-    prevProps: CanvasErrorBoundaryProps,
-    _prevState: CanvasErrorBoundaryState,
-  ): void {
-    if (prevProps.fileCode !== this.props.fileCode) {
-      // eslint-disable-next-line react/no-did-update-set-state
-      this.setState({ hasError: false })
-    }
-  }
-
-  componentDidCatch(error: unknown, errorInfo: React.ErrorInfo): void {
-    this.props.reportError(this.props.filePath, asErrorObject(error), errorInfo)
-  }
-
-  render(): React.ReactNode {
-    if (this.state.hasError) {
-      return null
-    } else {
-      return this.props.children
-    }
-  }
-}
